requirements.txt
# Core dependencies
opencv-python>=4.8.0
pyzbar>=0.1.9
Pillow>=10.0.0
requests>=2.31.0

# Web UI
streamlit>=1.28.0

# Database and caching
# SQLite is built into Python

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Desktop UI (uncomment if needed)
# PyQt5>=5.15.0
# tkinter is built into Python

# Development tools
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0

# Type hints
typing-extensions>=4.8.0
